Allow timdex readonly access to timdex prefixed indexes
Prep work for https://mitlibraries.atlassian.net/browse/DIP-225 in which we'll start using `timdex-production` instead of `production` since es is being treated as a shared resource and not only being used for timdex.

When the mario changes land and timdex ENV is updated, we should clean up the access to `"${module.shared.es_arn}/production*",`.