Refactor Geoweb app
This is largely a refactoring commit. The EC2 instance creation was
moved to an internal module since it was duplicated between Solr and
GeoServer. Most of the security groups were cleaned up as they were all
over the place.

The one major new change is the addition of the Cloudwatch cron job for
slingshot.