Remove broken bastion security group
The way this security group is being use does not work. The correct way
to provide access to a resource from the bastion host is to use the
security group associated with the bastion EC2 instance in an ingress
rule. This removes the broken security group and sets the output to the
correct security group.