Disable travis integration for now
We can't run our Makefile on this repo unless everything has been
switched over to 0.12. We'll need to rethink how we do this anyways
because 0.12 requires each config be initialized before the checks are
run.

Until we can get everything migrated we'll need to make sure we're
running fmt and validate locally.