Re-enabled Travis
This fixes the Makefile to work with 0.12. I've added a grep condition
to test for the config being upgraded to version 0.12, so the build
target will only be run on 0.12 configs.

make will take longer to run now since terraform init has to be called
on each config. I'd suggest running make with the -j option. This will
run the build target on each config in parallel.