Add archivesspace DNS entry for private dns zone
The way our DNS has been set up, none of the records from the public
hosted zone are available from the private hosted zone. If there are
cases when something from within the VPC needs to resolve a name in a
public hosted zone record it will need to be duplicated in the private
hosted zone.