Add r/w permissions to ASpace bucket
This adds r/w permissions to the s3 bucket being used by the OAI
harvester on the airflow worker role. It's not an ideal way to handle
this but until we can come up with a better way we'll just duplicate
whatever permissions the worker needs from the task role.