Fix cantaloupe deploy
There were some problems with the existing deploy. The bucket keys were
stored in clear text in the task definition. Also, the modules that were
used made it difficult to deploy a new version of the task definition.
These changes remove those modules and move the keys to SSM parameters.