Manage all users through Terraform
This brings all of our users into Terraform. I have switched from using
count to for_each to handle the user list. This is primarily because
with count the order of the listing matters and inserting a user anwhere
in the middle or the beginning of the list would lead to problems. A
side effect of this change is that we can't use the splat operator in
the output. I have just removed the output of users because I don't
believe this output is actually needed.

I've also imported our current group membership/policy structure so that
this too is managed by Terraform.